League of legends are added.
A option for modify league of legends was added in this version 0.8.
